Day 13, part 1
Got a major hint from the r/adventofcode subreddit. Even then, it took
me a good hour to relearn how to solve it using matrices. It's been
years since I've touched matrices and matrices are one of my weakest
math areas. I wrote the solution in Python first to get around parsing
it. After that it was a fairly straightforward translation to Go.